Move UI components to correct package
