+++ Corona-Warn-App/src/deviceForTesters/res/navigation/test_nav_graph.xml
@@ -169,7 +169,7 @@
         tools:layout="@layout/fragment_test_showstoredevents" />
     <fragment
         android:id="@+id/traceLocationOrganizerCategoriesFragment"
-        android:name="de.rki.coronawarnapp.eventregistration.events.ui.category.TraceLocationCategoryFragment"
+        android:name="de.rki.coronawarnapp.ui.eventregistration.organizer.category.TraceLocationCategoryFragment"
         android:label="TraceLocationCategoryFragment"
         tools:layout="@layout/trace_location_organizer_category_fragment" />
