--- Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/events/ui/category/adapter/separator/TraceLocationSeparatorItem.kt
@@ -1,7 +1,0 @@
-package de.rki.coronawarnapp.eventregistration.events.ui.category.adapter.separator
-
-import de.rki.coronawarnapp.eventregistration.events.ui.category.adapter.CategoryItem
-
-object TraceLocationSeparatorItem : CategoryItem {
-    override val stableId = this.hashCode().toLong()
-}
