+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/eventregistration/EventRegistrationUIModule.kt
@@ -2,14 +2,14 @@ package de.rki.coronawarnapp.ui.eventregistration
 
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
-import de.rki.coronawarnapp.eventregistration.events.ui.category.TraceLocationCategoryFragment
-import de.rki.coronawarnapp.eventregistration.events.ui.category.TraceLocationCategoryModule
 import de.rki.coronawarnapp.ui.eventregistration.attendee.checkins.CheckInsFragment
 import de.rki.coronawarnapp.ui.eventregistration.attendee.checkins.CheckInsModule
 import de.rki.coronawarnapp.ui.eventregistration.attendee.confirm.ConfirmCheckInFragment
 import de.rki.coronawarnapp.ui.eventregistration.attendee.confirm.ConfirmCheckInModule
 import de.rki.coronawarnapp.ui.eventregistration.attendee.scan.ScanCheckInQrCodeFragment
 import de.rki.coronawarnapp.ui.eventregistration.attendee.scan.ScanCheckInQrCodeModule
+import de.rki.coronawarnapp.ui.eventregistration.organizer.category.TraceLocationCategoryFragment
+import de.rki.coronawarnapp.ui.eventregistration.organizer.category.TraceLocationCategoryModule
 
 @Module
 internal abstract class EventRegistrationUIModule {
