+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/eventregistration/organizer/category/adapter/category/TraceLocationCategory.kt
@@ -1,10 +1,7 @@
-package de.rki.coronawarnapp.eventregistration.events.ui.category.adapter.category
+package de.rki.coronawarnapp.ui.eventregistration.organizer.category.adapter.category
 
 import androidx.annotation.StringRes
 import de.rki.coronawarnapp.R
-import de.rki.coronawarnapp.eventregistration.events.ui.category.adapter.CategoryItem
-import de.rki.coronawarnapp.eventregistration.events.ui.category.adapter.category.TraceLocationUIType.EVENT
-import de.rki.coronawarnapp.eventregistration.events.ui.category.adapter.category.TraceLocationUIType.LOCATION
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass.TraceLocationType.LOCATION_TYPE_PERMANENT_CRAFT
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass.TraceLocationType.LOCATION_TYPE_PERMANENT_EDUCATIONAL_INSTITUTION
@@ -18,6 +15,9 @@ import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass.TraceLocationType.LOCATION_TYPE_TEMPORARY_OTHER
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass.TraceLocationType.LOCATION_TYPE_TEMPORARY_PRIVATE_EVENT
 import de.rki.coronawarnapp.server.protocols.internal.pt.TraceLocationOuterClass.TraceLocationType.LOCATION_TYPE_TEMPORARY_WORSHIP_SERVICE
+import de.rki.coronawarnapp.ui.eventregistration.organizer.category.adapter.CategoryItem
+import de.rki.coronawarnapp.ui.eventregistration.organizer.category.adapter.category.TraceLocationUIType.EVENT
+import de.rki.coronawarnapp.ui.eventregistration.organizer.category.adapter.category.TraceLocationUIType.LOCATION
 
 data class TraceLocationCategory(
     val type: TraceLocationOuterClass.TraceLocationType,
