+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/eventregistration/organizer/category/adapter/header/TraceLocationHeaderItem.kt
@@ -1,7 +1,7 @@
-package de.rki.coronawarnapp.eventregistration.events.ui.category.adapter.header
+package de.rki.coronawarnapp.ui.eventregistration.organizer.category.adapter.header
 
 import androidx.annotation.StringRes
-import de.rki.coronawarnapp.eventregistration.events.ui.category.adapter.CategoryItem
+import de.rki.coronawarnapp.ui.eventregistration.organizer.category.adapter.CategoryItem
 
 data class TraceLocationHeaderItem(@StringRes val headerText: Int) : CategoryItem {
     override val stableId = this.hashCode().toLong()
