+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/eventregistration/organizer/category/adapter/separator/TraceLocationSeparatorItem.kt
@@ -1,0 +1,7 @@
+package de.rki.coronawarnapp.ui.eventregistration.organizer.category.adapter.separator
+
+import de.rki.coronawarnapp.ui.eventregistration.organizer.category.adapter.CategoryItem
+
+object TraceLocationSeparatorItem : CategoryItem {
+    override val stableId = this.hashCode().toLong()
+}
